% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mir_funs.R
\name{cor_make}
\alias{cor_make}
\title{Make mircoRNA feature correlations}
\usage{
cor_make(mi, m, cohort, tidy = TRUE)
}
\arguments{
\item{mi}{A tidy matrix of microRNA expression data.}

\item{m}{A tidy matrix of gene or protein expression data.}

\item{cohort}{A charachter string of the TCGA ID.}

\item{tidy}{A logical}
}
\value{
A tidy data.frame of four columns; miRBase ID, feature id, correlation and TCGA study name.
}
\description{
Calculate the pearson's correlation between microRNA and feature (gene/protein) matrices.
The function calls the R base cor function. More importantly the function prepare the input
matrices before the call. First, if matches the col.names/bcr. Second, remove duplicated entries.
Finally, it removes the microRNAs and features with expression values less than one. In addation,
The function returns a tidy data.frame rather than the raw matrix output of cor.
}
\examples{
# load required libraries
library(RTCGA.rnaseq)
library(RTCGA.miRNASeq)
library(RTCGA.RPPA)

# calculate correlation for genes
ACC.mi <- mirna_tidy(ACC.miRNASeq)
ACC.m <- mrna_tidy(ACC.rnaseq)
corr <- cor_make(ACC.mi, ACC.m, 'ACC')

# calculate correlation for proteins
ACC.mi <- mirna_tidy(ACC.miRNASeq)
ACC.r <- rppa_tidy(ACC.RPPA)
corr <- cor_make(ACC.mi, ACC.r, 'ACC')

}
