% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mir_funs.R
\name{get_targets}
\alias{get_targets}
\title{Get microRNA targets}
\usage{
get_targets(mirna, feature_type = "gene")
}
\arguments{
\item{mirna}{A charachter vector of the names of microRNAs of interes}

\item{feature_type}{A character string to choose the desired type of annotations. Takes one of 'gene' or 'protein'.}
}
\value{
A tidy data.frame of microRNA in one column 'mirna_base' and targets column 'feature'. When used with the 'protein' another
  column is added to distinguish the feature_types genes or proteins.
}
\description{
This function uses the targetscan database annotations to get the microRNA gene and protien tragets. For microRNA-gene targets
it is a straight forward subsetting the database and converting to ta tidy data.frame with the appropriate IDs. For microRNA-protein
targets, it first obtain the gene targets as explained and then change transform the IDs to the corresponding protein/antibody IDs
provided by the manufacturers of the used assays.
}
\examples{
mirna_base <- c("hsa-let-7a", "hsa-let-7b", "hsa-let-7c")
gene <- get_targets(mirna_base, 'gene')
ptn <- get_targets(mirna_base, 'protein')

}
