% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mir_funs.R
\name{mirna_tidy}
\alias{mirna_tidy}
\title{Tidy RTCGA.miRNASeq data.frame}
\usage{
mirna_tidy(mirna)
}
\arguments{
\item{mirna}{A RTCGA.miRNASeq data.frame}
}
\value{
A tidy expression matrix with microRNAs in the rows and patient samples in the columns.
}
\description{
This is a custom function to tidy up the RTCGA.miRNASeq data.frames. The aim is to strip out unnessecary data
and get to a typical bioconductor expression matrix. To achieve that, the function drops unneeded columns and rows,
extract the patient bcr and microRNA miRBase IDs and add them to the transposed matrix as column and row names.
}
\examples{
library(RTCGA.miRNASeq)
rtcga.df <- ACC.miRNASeq
ACC.mat <- mirna_tidy(rtcga.df)

}
