% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mir_funs.R
\name{mrna_tidy}
\alias{mrna_tidy}
\title{Tidy RTCGA.rnaseq data.frame}
\usage{
mrna_tidy(mrna)
}
\arguments{
\item{mrna}{A RTCGA.rnaseq data.frame}
}
\value{
A tidy expression matrix with genes in the rows and patient samples in the columns.
}
\description{
This is a custom function to tidy up the RTCGA.rnaseq data.frames. The aim is to strip out unnessecary data
and get to a typical bioconductor expression matrix. To achieve that, the function drops the first column,
extract the patient bcr and gene symbols and add them to the transposed matrix as column and row names.
}
\examples{
library(RTCGA.rnaseq)
rtcga.df <- ACC.rnaseq
ACC.mat <- mrna_tidy(rtcga.df)

}
