% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mir_funs.R
\name{rppa_tidy}
\alias{rppa_tidy}
\title{Tidy RTCGA.RPPA data.frame}
\usage{
rppa_tidy(rppa)
}
\arguments{
\item{rppa}{A RTCGA.RPPA data.frame}
}
\value{
A tidy expression matrix with proteins in the rows and patient samples in the columns.
}
\description{
This is a custom function to tidy up the RTCGA.RPPA data.frames. The aim is to strip out unnessecary data
and get to a typical bioconductor expression matrix. To achieve that, the function drops the first column,
extract the patient bcr and antibody ID and add them to the transposed matrix as column and row names.
}
\examples{
library(RTCGA.RPPA)
rtcga.df <- ACC.RPPA
ACC.mat <- rppa_tidy(rtcga.df)

}
