% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mir_funs.R
\name{sqlome_info}
\alias{sqlome_info}
\title{Extract TCGA info}
\usage{
sqlome_info()
}
\value{
A data.frame with Cohort, BCR and a columns with the number of cases in each assay
}
\description{
Extract TCGA info for mRNASeq, miRSeq and RPPA. The function calls the RTCGA::infoTCGA and
removes entries that doesn't have all three assays performed; mRNASeq, miRSeq and RPPA.
}
\examples{
sqlome_info

}
