% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_funs.R
\name{tf_read}
\alias{tf_read}
\title{Read transcription factor correlation data}
\usage{
tf_read(dir = "tmp/tf/")
}
\arguments{
\item{dir}{A character string of the directory containing the Cistrome Cancer files}
}
\value{
A named list of tidy data.frames. Each item of the list correspond to a transcription factor. The data.frame consist of
  a column containing the names of the features and a single column for each of the TCGA studies containing the correspoinding
  correlation value.
}
\description{
This is a custom function to read the local csv files obtained from Cistrome Cancer into a list of tidy data.frames.
First, the function constructs a vector of available files in the given directory and reads each of them to a data.frame.
Then, it drops the damaged files and names the items of the list with the corresponding transcription factor name.
}
